Wrapping js executable js obj as a function with metadata
In clojure, a function is just a function while in JavaScript, a
function is like any object which can have it own
properties.
We want map function in js <=> funciton in clojure so that we can
execute a js function just like a clojure function.
But we can't access its other properties after wrapping it in a
function, unless we use metadata, which is what we are trying to do
now.